Updated CharacterMovement; Created health/damage; Created Spawn; & some
changed characterMovement to use FloatDatas for speed. created health and damage. created spawn. death does not work yet but is in progress.

+++ Coding Samples/Assets/Scripts/FloatData.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using UnityEngine;
 
 [CreateAssetMenu]
-public class IntData : ScriptableObject
+public class FloatData : ScriptableObject
 {
-    public int myValue;
+    public float myValue;
 }
